Moved template to eskip package and adjusted docs.
